pyproject.toml
[project]
name = "blog-post"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "psycopg2-binary>=2.9.10",
]
